+++ pages/clients/index.vue
@@ -210,280 +210,48 @@
               </v-row>
 
               <v-row>
-                <v-col cols="12" sm="6" md="12">
-                  <v-divider></v-divider>
-                  <v-subheader>Contactos</v-subheader>
-                  <v-simple-table>
-                    <template v-slot:default>
-                      <thead>
-                        <tr>
-                          <th class="text-left">
-                            Principal
-                          </th>
-                          <th class="text-left">
-                            Tipo
-                          </th>
-                          <th class="text-left">
-                            Nombre
-                          </th>
-                          <th class="text-left">
-                            Apellidos
-                          </th>
-                          <th class="text-left">
-                            email
-                          </th>
-                          <th class="text-left">
-                            Teléfono Principal
-                          </th>
-                          <th class="text-left">
-                            Teléfono Secundario
-                          </th>
-                          <th class="text-left">
-                            Señas
-                          </th>
-                          <th class="text-left">
-                            <v-icon small class="mr-2" @click="createContact()">
-                              mdi-plus
-                            </v-icon>
-                          </th>
-                        </tr>
-                      </thead>
-                      <tbody>
-                        <tr v-for="item in contacts" :key="item.id">
-                          <td>
-                            <v-edit-dialog
-                              :return-value="item.isPrincipal"
-                              @save="saveContact()"
-                              @cancel="cancelContact"
-                              @open="openContact"
-                              @close="closeContact"
-                              large
-                              cancel-text="Cancelar"
-                              save-text="Guardar"
-                              persistent
-                            > 
-                              <v-icon small class="mr-2" v-if="item.isPrincipal">
-                                mdi-check
-                              </v-icon>
-                              <v-icon small class="mr-2" v-if="!item.isPrincipal">
-                                mdi-close
-                              </v-icon>
-                              <template v-slot:input>
-                                <v-container
-                                  class="px-0"
-                                  fluid
-                                >
-                                  <v-checkbox
-                                    :value="item.isPrincipal"
-                                    @change="updateContact(item, $event, 'isPrincipal')"
-                                    :label="`Es Principal: ${item.isPrincipal ? 'Sí' : 'No'}`"
-                                  ></v-checkbox>
-                                </v-container>
-                              </template>
-                            </v-edit-dialog>
-                          </td>
-                          <td>
-                            <v-edit-dialog
-                              :return-value="item.type"
-                              @save="saveContact()"
-                              @cancel="cancelContact"
-                              @open="openContact"
-                              @close="closeContact"
-                              large
-                              cancel-text="Cancelar"
-                              save-text="Guardar"
-                              persistent
-                            > {{ item.type }}
-                              <template v-slot:input>
-                                <ValidationProvider v-slot="{ errors }" name="Tipo" rules="required">
-                                  <v-text-field
-                                    :value="item.type"
-                                    @change="updateContact(item, $event, 'type')"
-                                    label="Tipo *"
-                                    single-line
-                                    :error-messages="errors"
-                                    required
-                                  ></v-text-field>
-                                </ValidationProvider>
-                              </template>
-                            </v-edit-dialog>
-                          </td>
-                          <td>
-                            <v-edit-dialog
-                              :return-value="item.name"
-                              @save="saveContact()"
-                              @cancel="cancelContact"
-                              @open="openContact"
-                              @close="closeContact"
-                              large
-                              cancel-text="Cancelar"
-                              save-text="Guardar"
-                              persistent
-                            > {{ item.name }}
-                              <template v-slot:input>
-                                <ValidationProvider v-slot="{ errors }" name="Nombre" rules="required">
-                                  <v-text-field
-                                    :value="item.name"
-                                    @change="updateContact(item, $event, 'name')"
-                                    label="Nombre *"
-                                    single-line
-                                    :error-messages="errors"
-                                    required
-                                  ></v-text-field>
-                                </ValidationProvider>
-                              </template>
-                            </v-edit-dialog>
-                          </td>
-                          <td>
-                            <v-edit-dialog
-                              :return-value="item.lastName"
-                              @save="saveContact()"
-                              @cancel="cancelContact"
-                              @open="openContact"
-                              @close="closeContact"
-                              large
-                              cancel-text="Cancelar"
-                              save-text="Guardar"
-                              persistent
-                            > {{ item.lastName }}
-                              <template v-slot:input>
-                                <ValidationProvider v-slot="{ errors }" name="Apellidos" rules="required">
-                                  <v-text-field
-                                    :value="item.lastName"
-                                    @change="updateContact(item, $event, 'lastName')"
-                                    label="Apellidos *"
-                                    single-line
-                                    :error-messages="errors"
-                                    required
-                                  ></v-text-field>
-                                </ValidationProvider>
-                              </template>
-                            </v-edit-dialog>
-                          </td>
-                          <td>
-                            <v-edit-dialog
-                              :return-value="item.email"
-                              @save="saveContact()"
-                              @cancel="cancelContact"
-                              @open="openContact"
-                              @close="closeContact"
-                              large
-                              cancel-text="Cancelar"
-                              save-text="Guardar"
-                              persistent
-                            > {{ item.email }}
-                              <template v-slot:input>
-                                <ValidationProvider v-slot="{ errors }" name="Email" rules="required|email">
-                                  <v-text-field
-                                    :value="item.email"
-                                    @change="updateContact(item, $event, 'email')"
-                                    label="Email *"
-                                    single-line
-                                    :error-messages="errors"
-                                    required
-                                  ></v-text-field>
-                                </ValidationProvider>
-                              </template>
-                            </v-edit-dialog>
-                          </td>
 
-                          <td>
-                            <v-edit-dialog
-                              :return-value="item.mobile"
-                              @save="saveContact()"
-                              @cancel="cancelContact"
-                              @open="openContact"
-                              @close="closeContact"
-                              large
-                              cancel-text="Cancelar"
-                              save-text="Guardar"
-                              persistent
-                            > {{ item.mobile }}
-                              <template v-slot:input>
-                                <ValidationProvider v-slot="{ errors }" name="Teléfono Principal" rules="required|digits:8">
-                                  <v-text-field
-                                    :value="item.mobile"
-                                    @change="updateContact(item, $event, 'mobile')"
-                                    label="Teléfono Principal *"
-                                    single-line
-                                    :type="'number'"
-                                    hint="8 números"
-                                    required
-                                    :error-messages="errors"
-                                  ></v-text-field>
-                                </ValidationProvider>
-                              </template>
-                            </v-edit-dialog>
-                          </td>
+                <template>
 
-                          <td>
-                            <v-edit-dialog
-                              :return-value="item.phone"
-                              @save="saveContact()"
-                              @cancel="cancelContact"
-                              @open="openContact"
-                              @close="closeContact"
-                              large
-                              cancel-text="Cancelar"
-                              save-text="Guardar"
-                              persistent
-                            > {{ item.phone }}
-                              <template v-slot:input>
-                                <ValidationProvider v-slot="{ errors }" name="Teléfono Secundario" rules="digits:8">
-                                  <v-text-field
-                                    :value="item.phone"
-                                    @change="updateContact(item, $event, 'phone')"
-                                    label="Teléfono Secundario"
-                                    single-line
-                                    :type="'number'"
-                                    hint="8 números"
-                                    :error-messages="errors"
-                                  ></v-text-field>
-                                </ValidationProvider>
-                              </template>
-                            </v-edit-dialog>
-                          </td>
+                  <v-col cols="12" sm="6" md="12">
+
+                    <v-expansion-panels
+                      v-model="panel"
+                      multiple
+                    >
+
+                      <v-expansion-panel>
+                        <v-expansion-panel-header>
+                          Fincas
+                        </v-expansion-panel-header>
+                        <v-expansion-panel-content>
                           
-                          <td>
-                            <v-edit-dialog
-                              :return-value="item.signs"
-                              @save="saveContact()"
-                              @cancel="cancelContact"
-                              @open="openContact"
-                              @close="closeContact"
-                              large
-                              cancel-text="Cancelar"
-                              save-text="Guardar"
-                              persistent
-                            > {{ item.signs }}
-                              <template v-slot:input>
-                                <ValidationProvider v-slot="{ errors }" name="Señas">
-                                  <v-text-field
-                                    :value="item.signs"
-                                    @change="updateContact(item, $event, 'signs')"
-                                    label="Señas *"
-                                    single-line
-                                    :error-messages="errors"
-                                    required
-                                  ></v-text-field>
-                                </ValidationProvider>
-                              </template>
-                            </v-edit-dialog>
-                          </td>
-                          <td>
-                            <v-icon small @click="deleteContact(item.id)">
-                              mdi-delete
-                            </v-icon>
-                          </td>
-                        </tr>
-                      </tbody>
-                    </template>
-                  </v-simple-table>
-                </v-col>
+                        </v-expansion-panel-content>
+                      </v-expansion-panel>
+
+                      <v-expansion-panel>
+                        <v-expansion-panel-header>
+                          Conctactos
+                        </v-expansion-panel-header>
+                        <v-expansion-panel-content>
+                          
+                          <template>
+                            <contacts-vue :currentClient="currentClient"></contacts-vue>
+                          </template>
+
+                        </v-expansion-panel-content>
+                      </v-expansion-panel>
+
+                    </v-expansion-panels>
+
+                  </v-col>
+
+                </template>
+
               </v-row>
 
             </v-container>
-            <small>*campos requeridos</small>
+
           </v-card-text>
           <v-card-actions>
             <v-spacer></v-spacer>
@@ -614,8 +382,13 @@
 </template>
 
 <script>
+import ContactsVue from './contacts.vue'
+
 export default {
   name: "clients",
+  components: {
+    ContactsVue
+  },
   data: () => ({
     clientDialog: false,
     deleteClientDialog: false,
@@ -669,6 +442,7 @@ export default {
     actionSuccess: false,
     loaderActive: false,
     isEdition: false,
+    panel: []
   }),
   async fetch() {
     this.loaderActive = true;
@@ -686,7 +460,6 @@ export default {
         cantonId: this.client.canton
       });
     } catch (error) {
-      console.log(error)
       this.activateSnackbar("Obteniendo la información " + error, false);
     }
 
@@ -695,9 +468,6 @@ export default {
   computed: {
     clients(){
       return this.$store.getters['clients/clients'];
-    },
-    contacts(){
-      return this.$store.getters['contacts/contacts'];
     },
     provincias(){
       return this.$store.getters['locations/provincias'];
@@ -744,7 +514,6 @@ export default {
 
       this.$fetch();
     },
-
     openUpdateClientDialog(data) {
       this.currentClient = data;
       this.isEdition = true;
@@ -767,26 +536,21 @@ export default {
         distrito: data.distrito,
         address: data.address
       };
-
       this.$fetch();
     },
-
     closeClientDialog() {
       this.clientDialog = false;
       this.currentClient = null;
       this.$refs.observer.reset();
     },
-
     openDeleteClientDialog(item) {
       this.deleteClientDialog = true;
       this.currentClient = item;
     },
-
     closeDeleteClientDialog() {
       this.deleteClientDialog = false;
       this.currentClient = null;
     },
-
     async createClient() {
       const isValid = await this.$refs.observer.validate();
 
@@ -828,7 +592,6 @@ export default {
           });
       }
     },
-
     async updateClient() {
       const isValid = await this.$refs.observer.validate();
 
@@ -872,7 +635,6 @@ export default {
           });
       }
     },
-
     async deleteClient() {
       this.loaderActive = true;
 
@@ -894,7 +656,6 @@ export default {
           this.loaderActive = false;
         });
     },
-
     activateSnackbar(message, success) {
       this.snackbar.text = message;
       this.snackbar.visible = true;
@@ -909,103 +670,11 @@ export default {
         this.snackbar.title = "Error";
       }
     },
-
-    async updateContact(contact, newValue, field) {
-      const isValid = await this.$refs.observer.validate();
-
-      if (isValid) {
-        let newContact = JSON.parse(JSON.stringify(contact));
-        newContact[field] = newValue;
-
-        await this.$store.dispatch('contacts/updateContact', {
-          newContact: newContact
-        });
-
-        this.$refs.observer.reset();
-      }
-    },
-
-    async saveContact() {
-      await this.$store.dispatch('contacts/saveContact', {
-        currentContact: this.$store.getters['contacts/contact']
-      });
-
-      this.$fetch();
-    },
-
-    cancelContact() {
-      this.$refs.observer.reset();
-    },
-
-    openContact() {
-      this.$refs.observer.reset();
-    },
-
-    closeContact() {
-      this.$refs.observer.reset();
-    },
-
-    async createContact() {
-      this.loaderActive = true;
-
-      if(this.currentClient){
-        await this.$fire.firestore
-        .collection("contacts")
-        .add({
-          email: 'editar@editar.com',
-          isPrincipal: '',
-          lastName: 'Seleccione...',
-          mobile: '87654321',
-          name: 'Seleccione...',
-          phone: '45678909',
-          signs: 'Seleccione...',
-          type: 'Seleccione...',
-          clientId: this.currentClient.id,
-          timestamp: new Date()
-        })
-        .then(() => {
-          this.$fetch();
-          this.activateSnackbar("Contacto creado, proceda a editarlo", true);
-          this.loaderActive = false;
-        })
-        .catch((error) => {
-          console.error("Error writing document: ", error);
-          this.activateSnackbar("Error creando contacto", false);
-
-          this.loaderActive = false;
-        });
-      } else {
-        this.activateSnackbar("Para poder crear un contacto debe crear el cliente primero", false);
-
-        this.loaderActive = false;
-      }
-    },
-
-    async deleteContact(contactId) {
-      this.loaderActive = true;
-
-      await this.$fire.firestore
-        .collection("contacts")
-        .doc(contactId)
-        .delete()
-        .then(() => {
-          this.$fetch();
-          this.activateSnackbar("Contacto borrado correctamente", true);
-          this.loaderActive = false;
-        })
-        .catch((error) => {
-          console.error("Error removing document: ", error);
-          this.activateSnackbar("Error borrando contacto", false);
-
-          this.loaderActive = false;
-        });
-    },
-
     getClientTypeText(type){
       return this.clientTypeList.filter((item) => {
         return item.value == type;
       })[0].text;
-    },
+    }
   }
 };
 </script>
